Move railgun projectile registration to separate class
